Update to latest React Native Version

+++ ReactAndroid/src/main/java/com/facebook/react/views/scroll/ReactHorizontalScrollViewManager.java
@@ -1,10 +1,8 @@
 /**
  * Copyright (c) 2015-present, Facebook, Inc.
- * All rights reserved.
  *
- * This source code is licensed under the BSD-style license found in the
- * LICENSE file in the root directory of this source tree. An additional grant
- * of patent rights can be found in the PATENTS file in the same directory.
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
  */
 
 package com.facebook.react.views.scroll;
@@ -73,9 +71,10 @@ public class ReactHorizontalScrollViewManager
   }
 
   @ReactProp(name = "snapToInterval")
-  public void setSnapToInterval(ReactHorizontalScrollView view, int snapToInterval) {
+  public void setSnapToInterval(ReactHorizontalScrollView view, float snapToInterval) {
+    // snapToInterval needs to be exposed as a float because of the Javascript interface.
     DisplayMetrics screenDisplayMetrics = DisplayMetricsHolder.getScreenDisplayMetrics();
-    view.setSnapInterval((int)(snapToInterval * screenDisplayMetrics.density));
+    view.setSnapInterval((int) (snapToInterval * screenDisplayMetrics.density));
   }
 
   @ReactProp(name = ReactClippingViewGroupHelper.PROP_REMOVE_CLIPPED_SUBVIEWS)
